+++ packages/api/src/repositories/types.ts
@@ -21,19 +21,15 @@ export type {
   Review,
 } from '../stores'
 export type { DashboardStats } from '@kuruma/shared/types/stats'
-export type { OperatorOverview } from '@kuruma/shared/types/overview'
-export type { FleetVehicleOverview, FleetBookingSummary } from '@kuruma/shared/types/fleet'
 export type { VehicleDetail } from '@kuruma/shared/types/vehicle-detail'
 export type { Customer, CustomerSort, CustomerWithBookings } from '@kuruma/shared/types/customer'
 
 import type { CoordinateSource } from '@kuruma/shared/db/schema'
 import type { Customer, CustomerSort, CustomerWithBookings } from '@kuruma/shared/types/customer'
-import type { FleetVehicleOverview } from '@kuruma/shared/types/fleet'
 import type {
   OperatorBookingCounts,
   VehicleComplianceCounts,
 } from '@kuruma/shared/types/operator-summary'
-import type { OperatorOverview } from '@kuruma/shared/types/overview'
 import type { DashboardStats } from '@kuruma/shared/types/stats'
 import type { VehicleDetail } from '@kuruma/shared/types/vehicle-detail'
 import type { OperatorRole } from '@kuruma/shared/validators/provider-invite'
@@ -350,25 +346,6 @@ export interface VehicleRepository {
   removePhotoByUrl(ctx: CallerContext, id: string, url: string): Promise<Vehicle | undefined>
 }
 
-// Aggregated read for the owner-facing /manage/vehicles list. Enriches
-// each vehicle with utilization %, booking count, and current/next
-// booking state. Computed per-request — NOT denormalized into the
-// vehicles table. See issue #52 and @kuruma/shared/types/fleet.
-//
-// Split from VehicleRepository because it reads across multiple tables
-// (vehicles + bookings + users.name) — following the same boundary as
-// AvailabilityRepository, which also reads vehicles + bookings.
-export interface FleetOverviewRepository {
-  // `now` is injected so time cutoffs live in callers; tests pass a fixed Date.
-  // `ctx` scopes the read (#594); `operatorId` (#407 slice 4, bypass-gated)
-  // narrows a bypass caller to one operator — see narrowReadToOperator.
-  findFleetOverview(
-    ctx: CallerContext,
-    now: Date,
-    operatorId?: string,
-  ): Promise<FleetVehicleOverview[]>
-}
-
 /**
  * #396: UserRepository is intentionally NOT operator-scoped — it takes no
  * CallerContext. This stays safe because every ingress already blocks
@@ -522,18 +499,6 @@ export interface BookingRepository {
 
 export interface StatsRepository {
   getDashboardStats(): Promise<DashboardStats>
-}
-
-/**
- * Operator-scoped dashboard counts (#524). `ctx` decides the tenant scope via
- * {@link bookingReadScope}: bypass roles aggregate across all operators, an
- * OPERATOR_* caller sees only its own tenant, and an operator missing its
- * operatorId fails closed to zeros (mirrors how its own bookings list behaves).
- * `now` is injected so "upcoming" is deterministic in tests.
- */
-export interface OverviewRepository {
-  // `operatorId` (#407 slice 4, bypass-gated): narrows a bypass caller — see narrowReadToOperator.
-  getOperatorOverview(ctx: CallerContext, now: Date, operatorId?: string): Promise<OperatorOverview>
 }
 
 /**
@@ -740,3 +705,15 @@ export type {
 // Reviews bounded-context data access (#1067 slice 1) lives in its own module;
 // re-exported for callers (mirrors the payment/consent split above).
 export type { NewReview, ReviewEdit, ReviewRepository } from './types-review'
+
+// Dashboard overview + fleet-overview aggregate reads (#1265) live in their own
+// module to keep this barrel under the file-size cap; re-exported for callers.
+// The projection DTOs (OperatorOverview / FleetVehicleOverview / FleetBookingSummary)
+// ride along so the barrel keeps its historical surface.
+export type {
+  FleetBookingSummary,
+  FleetOverviewRepository,
+  FleetVehicleOverview,
+  OperatorOverview,
+  OverviewRepository,
+} from './types-overview'
